lab_6/pyproject.toml
[project]
name = "lab-6"
version = "0.1.0"
description = "Kompresja obrazów TGA (Sub-band + DPCM + LBG)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26.0",
    "pillow>=10.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]